% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_iea_mw.R
\name{add_iea_mw_psut}
\alias{add_iea_mw_psut}
\title{Title}
\usage{
add_iea_mw_psut(
  iea_psut,
  mw_psut,
  countries,
  R = IEATools::psut_cols$R,
  U = IEATools::psut_cols$U,
  V = IEATools::psut_cols$V,
  Y = IEATools::psut_cols$Y,
  U_feed = IEATools::psut_cols$U_feed,
  U_eiou = IEATools::psut_cols$U_eiou,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  r_eiou = IEATools::psut_cols$r_eiou
)
}
\arguments{
\item{iea_psut}{An IEA PSUT data frame.}

\item{mw_psut}{A muscle work PSUT data frame.}

\item{R}{The name of the column of \code{R} matrices. Default is \code{IEATools::psut_cols$R}.}

\item{U}{The name of the column of \code{U} matrices. Default is \code{IEATools::psut_cols$U}.}

\item{V}{The name of the column of \code{V} matrices. Default is \code{IEATools::psut_cols$V}.}

\item{Y}{The name of the column of \code{Y} matrices. Default is \code{IEATools::psut_cols$Y}.}
}
\value{
A data frame of summed matrices.
}
\description{
Title
}
