% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_iea_mw.R
\name{aggcountries_mw_to_iea}
\alias{aggcountries_mw_to_iea}
\title{Aggregate ILO and FAO country data into IEA country groupings}
\usage{
aggcountries_mw_to_iea(
  mw_df,
  exemplar_table_path = PFUSetup::get_abs_paths()$exemplar_table_path,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  unit = IEATools::iea_cols$unit,
  e_dot = IEATools::iea_cols$e_dot,
  agg_code_col = PFUDatabase::exemplar_names$agg_code_col,
  region_code = PFUDatabase::exemplar_names$region_code,
  exemplar_country = PFUDatabase::exemplar_names$exemplar_country,
  species = MWTools::mw_constants$species,
  stage_col = MWTools::mw_constants$stage_col,
  sector_col = MWTools::mw_constants$sector_col
)
}
\arguments{
\item{mw_df}{A data frame containing raw animal muscle work or human muscle
work data. Usually retrieved from the \code{AMWPFUDataRaw} and
\code{HMWPFUDataRaw} targets.}

\item{exemplar_table_path}{See \code{PFUSetup::get_abs_paths()}.}

\item{country, year, unit, e_dot}{See \code{IEATools::iea_cols}.}

\item{agg_code_col, region_code, exemplar_country}{See \code{PFUDatabase::exemplar_names}.}

\item{species, stage_col, sector_col}{See \code{MWTools::mw_constants}.}
}
\description{
This function reads an exemplar table containing a countries country code and
the country code associated with the territory that the IEA data was recorded
in for the years 1960 - 2020. A country as defined by it's territorial
boundaries in 2020 may have had it's energy statistics recorded in any number
of regions, over any number of years.
}
