% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_countries}
\alias{all_countries}
\title{All countries and selected country groups}
\format{
A string list with 175 entries.
}
\usage{
all_countries
}
\description{
All individual countries and selected country groups from the IEA's World Extended
energy balances 2021.
}
\examples{
all_countries
}
\keyword{datasets}
