% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{alloc_plots_df}
\alias{alloc_plots_df}
\title{Create allocation graphs in a data frame}
\usage{
alloc_plots_df(
  .df,
  countries,
  data_col = "Data",
  plots = "Plots",
  country = IEATools::iea_cols$country,
  ef_product = IEATools::template_cols$ef_product,
  destination = IEATools::template_cols$destination,
  quantity = IEATools::template_cols$quantity,
  c_source = IEATools::template_cols$c_source,
  year = IEATools::iea_cols$year,
  .values = IEATools::template_cols$.values,
  machine = IEATools::template_cols$machine,
  eu_product = IEATools::template_cols$eu_product
)
}
\arguments{
\item{.df}{The completed allocation tables data frame. Default is \code{drake::readd(completed_allocation_tables_target, path = cache_path, character_only = TRUE)}.}

\item{countries}{The countries for which allocation plots are to be created.}

\item{data_col}{The name of the output column containing nested data for allocation graphs. Default is "Data".}

\item{plots}{The name of the output column containing allocation graphs. Default is "Plots".}

\item{country}{See \code{IEATools::iea_cols}.}

\item{ef_product, destination, quantity, c_source}{See \code{IEATools::template_cols}.}

\item{year}{See \code{IEATools::iea_cols}. Passed to \code{alloc_graph()}.}

\item{.values, machine, eu_product}{See \code{IEATools::template_cols}. Passed to \code{alloc_graph()}.}
}
\value{
A data frame containing a list column of \code{ggplot2} allocation graphs.
}
\description{
This function adds a column of \code{ggplot2} graphs to a completed allocation tables data frame.
The graphs are stored in a list column named \code{plots}.
}
\details{
The data frame is grouped by all variables needed to create the allocation graph, specifically
\code{country}, \code{ef_product}, and \code{destination} and nested prior to making the graphs, namely
\code{machine}, \code{eu_product}, \code{quantity}, \code{year}, \code{.values}, and \code{c_source},
meaning that one allocation graph is constructed for each combination of those variables.
}
\examples{
# Make a simple data frame with the expected structure.
alloc_table <- tibble::tribble(~Country, ~Method, ~Energy.type, ~Year, ~Ef.product, ~Destination,
                               ~.values, ~Machine, ~Quantity, ~Eu.product, ~C.source,
                               "GHA", "PCM", "E", 1971, "Gasoline", "Transport",
                               0.5, "Cars", "C_1 [\%]", "MD", "World",
                               "GHA", "PCM", "E", 1971, "Gasoline", "Transport",
                               0.5, "Trucks", "C_2 [\%]", "MD", "World",
                               "GHA", "PCM", "E", 2020, "Gasoline", "Transport",
                               0.2, "Cars", "C_1 [\%]", "MD", "World",
                               "GHA", "PCM", "E", 2020, "Gasoline", "Transport",
                               0.8, "Trucks", "C_2 [\%]", "MD", "World",
                               "ZAF", "PCM", "E", 1971, "Gasoline", "Transport",
                               0.5, "Cars", "C_1 [\%]", "MD", "World",
                               "ZAF", "PCM", "E", 1971, "Gasoline", "Transport",
                               0.5, "Trucks", "C_2 [\%]", "MD", "World",
                               "ZAF", "PCM", "E", 2020, "Gasoline", "Transport",
                               0.3, "Cars", "C_1 [\%]", "MD", "World",
                               "ZAF", "PCM", "E", 2020, "Gasoline", "Transport",
                               0.7, "Trucks", "C_2 [\%]", "MD", "World")
alloc_plots_df(alloc_table, c("GHA", "ZAF"))
}
