% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_iea_mw.R
\name{build_psut_dataframe}
\alias{build_psut_dataframe}
\title{Build the final PSUT data frame}
\usage{
build_psut_dataframe(
  psutiea = NULL,
  psutmw = NULL,
  psutieamw = NULL,
  ieamw_colname = PFUDatabase::ieamw_cols$ieamw,
  iea = PFUDatabase::ieamw_cols$iea,
  mw = PFUDatabase::ieamw_cols$mw,
  both = PFUDatabase::ieamw_cols$both,
  year = IEATools::iea_cols$year
)
}
\arguments{
\item{psutiea}{A PSUT data frame of IEA data. Default is \code{NULL.}}

\item{psutmw}{A PSUT data frame of muscle work data. Default is \code{NULL.}}

\item{psutieamw}{A PSUT data frame of combined IEA and MW data. Default is \code{NULL.}}

\item{ieamw_colname}{The name of the column that identifies data source (IEA, MW, or both).
Default is \code{PFUDatabase::ieamw_cols$ieamw}.}

\item{iea}{The string that identifies ECC data are from the IEA only.
Default is \code{PFUDatabase::ieamw_cols$iea}.}

\item{mw}{The string that identifies ECC data are for muscle work only.
Default is \code{PFUDatabase::ieamw_cols$mw}.}

\item{both}{The string that identifies ECC data are for both IEA and muscle work.
Default is \code{PFUDatabase::ieamw_cols$both}.}

\item{year}{The string name of the year column.
Default is \code{IEATools::iea_cols$year}.}
}
\value{
A data frame with \code{PSUTIEA}, \code{PSUTMW}, and \code{PSUTIEAMW} \code{rbind()}ed together,
and a new column (\code{IEAMW_colname}) that distinguishes among them.
}
\description{
Combines PSUT descriptions based on IEA data exclusively, muscle work data exclusively,
and summed IEA and MW data.
}
