% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_to_useful.R
\name{calc_C_mats}
\alias{calc_C_mats}
\title{Add allocation matrices to a data frame}
\usage{
calc_C_mats(
  completed_allocation_tables,
  countries,
  matrix_class = c("matrix", "Matrix"),
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  c_source = IEATools::template_cols$c_source,
  .values = IEATools::template_cols$.values,
  C_Y = IEATools::template_cols$C_Y,
  C_EIOU = IEATools::template_cols$C_eiou
)
}
\arguments{
\item{completed_allocation_tables}{The completed allocation tables from which allocation (\code{C}) matrices should be created.
This data frame is most likely to be the \code{CompletedAllocationTables} target.}

\item{countries}{The countries for which \code{C} matrices should be formed.}

\item{matrix_class}{The type of matrix that should be produced.
One of "matrix" (the default and not sparse) or "Matrix" (which may be sparse).}

\item{country, year}{See \code{IEATools::iea_cols}.}

\item{c_source, .values, C_Y, C_EIOU}{See \code{IEATools::template_cols}.}
}
\value{
A data frame with \code{C_Y} and \code{C_EIOU} columns containing allocation matrices.
}
\description{
This function adds allocation matrices (\code{C_Y} and \code{C_EIOU}) to the previously-created
\code{CompletedAllocationTables} target.
}
