% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiencies.R
\name{calc_C_mats_agg}
\alias{calc_C_mats_agg}
\title{Calculates the aggregated C matrices}
\usage{
calc_C_mats_agg(
  C_mats,
  psut_iea,
  C_EIOU = "C_EIOU",
  C_Y = "C_Y",
  Y = "Y",
  U_EIOU = "U_EIOU",
  C_EIOU_agg = "C_EIOU_agg",
  C_Y_agg = "C_Y_agg",
  C_EIOU_Y_agg = "C_EIOU_Y_agg",
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year
)
}
\arguments{
\item{C_mats}{A data frame containing the C matrices.}

\item{psut_iea}{A data frame containing IEA data at the final stage, in energy terms, in a PSUT format.}

\item{C_EIOU}{The name of the column containing the C_EIOU matrix in the input data frame.}

\item{C_Y}{The name of the column containing the C_Y matrix in the input data frame.}

\item{Y}{The name of the column containing the Y matrix in the input data frame.}

\item{U_EIOU}{The name of the column containing the U_EIOU matrix in the input data frame.}

\item{C_EIOU_agg}{The name of the column containing the C_EIOU aggregated matrix in the output data frame.}

\item{C_Y_agg}{The name of the column containing the C_Y aggregated matrix in the output data frame.}

\item{C_EIOU_Y_agg}{The name of the column containing the C_EIOU_Y aggregated matrix in the output data frame.}

\item{country, method, energy_type, last_stage, year}{See \code{IEATools::iea_cols}.}
}
\value{
A data frame containing the aggregated C matrices for EIOU, Y and EIOU and Y together.
}
\description{
Calculates the aggregated C matrices
}
