% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exiobase.R
\name{calc_Ef_to_Eloss_exiobase}
\alias{calc_Ef_to_Eloss_exiobase}
\title{Calculates the final energy to energy losses multipliers}
\usage{
calc_Ef_to_Eloss_exiobase(ExiobaseEftoEuMultipliers_df)
}
\arguments{
\item{ExiobaseEftoEuMultipliers_df}{The data frame of final energy to useful energy multipliers previously calculated}
}
\value{
A data frame of the final energy to energy losses multipliers
}
\description{
Calculates the final energy to energy losses multipliers
}
