% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exiobase.R
\name{calc_Ef_to_Eu_exiobase}
\alias{calc_Ef_to_Eu_exiobase}
\title{Calculates the final energy to useful energy multipliers}
\usage{
calc_Ef_to_Eu_exiobase(
  EtafuYEIOU_mats,
  years_exiobase,
  full_list_exiobase_flows,
  country_concordance_table_df
)
}
\arguments{
\item{years_exiobase}{The years for which the coefficients are provided to the Exiobase team}

\item{full_list_exiobase_flows}{The full list of energy flows used in the Exiobase pipeline led by KR}

\item{country_concordance_table_df}{A data frame containing the country concordance table}

\item{EtafuYEIOU_df}{The input data frame containing matrices with all the efficiencies by final demand sector and energy industry}
}
\value{
A dataframe of the final energy to useful energy multipliers
}
\description{
Calculates the final energy to useful energy multipliers
}
