% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exiobase.R
\name{calc_Ef_to_Xloss_exiobase}
\alias{calc_Ef_to_Xloss_exiobase}
\title{Calculates the final energy to exergy losses multipliers}
\usage{
calc_Ef_to_Xloss_exiobase(ExiobaseEftoXuMultipliers_df)
}
\arguments{
\item{ExiobaseEftoXuMultipliers_df}{The data frame of final energy to useful exergy multipliers previously calculated}
}
\value{
A data frame of the final energy to exergy losses multipliers
}
\description{
Calculates the final energy to exergy losses multipliers
}
