% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exiobase.R
\name{calc_Ef_to_Xu_exiobase}
\alias{calc_Ef_to_Xu_exiobase}
\title{Calculates the final energy to useful exergy multipliers}
\usage{
calc_Ef_to_Xu_exiobase(
  EtafuYEIOU_mats,
  phi_vecs,
  eta_fu_phi_Y_EIOU_agg,
  years_exiobase,
  full_list_exiobase_flows,
  country_concordance_table_df,
  useful_energy_flow = "Useful.energy.flow",
  exiobase_flow = "Exiobase.Flow",
  pfu_code = "PFU.code",
  pfu_flow = "PFU.flow",
  iea_country_name = "IEA.country.name",
  iea_country_name_accented = "IEA.name.accented",
  phi = "phi",
  matnames = "matnames",
  colnames = "colnames",
  matvals = "matvals",
  rowtypes = "rowtypes",
  coltypes = "coltypes",
  energy_type_E = IEATools::energy_types$e,
  country = IEATools::iea_cols$country,
  product = IEATools::iea_cols$product,
  year = IEATools::iea_cols$year,
  flow = IEATools::iea_cols$flow,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  eta_p_eiou = "eta_p_eiou",
  eta_p_eiou_y = "eta_p_eiou_y",
  eta_fu_Y_E = "eta_fu_Y_E",
  eta_fu_EIOU_E = "eta_fu_EIOU_E",
  eta_phi_p_eiou = "eta_phi_p_eiou",
  eta_phi_p_eiou_y = "eta_phi_p_eiou_y",
  phi_eta_X = "phi_eta_X",
  eta = "eta"
)
}
\arguments{
\item{EtafuYEIOU_mats}{The input data frame containing matrices with all the efficiencies by final demand sector and energy industry}

\item{phi_vecs}{A data frame of phi (exergy-to-energy ratio) coefficients.}

\item{eta_fu_phi_Y_EIOU_agg}{A data frame containing matrices with the economy-wide efficiencies time phi values by energy product}

\item{years_exiobase}{The years for which the coefficients are provided to the Exiobase team}

\item{full_list_exiobase_flows}{The full list of energy flows used in the Exiobase pipeline led by KR}

\item{country_concordance_table_df}{A data frame containing the country concordance table}

\item{useful_energy_flow}{The name of the column stating whether a flow is a final energy flow or not}

\item{exiobase_flow}{The name of the column stating the name of the Exiobase flow}

\item{pfu_code}{The name of the column containing the PFU country name}

\item{iea_country_name}{The name of the column containing the IEA country name}

\item{phi}{The name of the column containing the phi values}

\item{matnames}{The name of the column containing matrices names after unpacking the matrices}

\item{matvals}{The name of the column containing matrices values after unpacking the matrices}

\item{rowtypes}{The name of the column containing the matrices row types names after unpacking the matrices}

\item{coltypes}{The name of the column containing matrices column types after unpacking the matrices}

\item{energy_type_E}{The letter standing for energy as energy type.}

\item{country, year, product, flow, method, energy_type, last_stage}{See \code{IEATools::iea_cols}.}

\item{eta_p_eiou}{The name of the column containing the efficiencies.}

\item{eta_fu_Y_E}{The name of the the column containing the efficiencies of products when used as part of final demand.}

\item{eta_fu_EIOU_E}{The name of the column containing the efficiencies of products when used as part of the EIOU matrix.}

\item{eta_phi_p_eiou}{The name of the column with the matrices containing the efficiencies, multiplied by the phi values, of products when used as part of the EIOU matrix.}

\item{eta_phi_p_eiou_y}{The name of the column with the matrices containing the efficiencies, multiplied by the phi values, of products when used as part of the Y matrix.}

\item{phi_eta_X}{The name of the column containing the efficiencies, multiplied by the phi values, of products when used as part of the EIOU matrix.}

\item{eta}{The name of the column containing the efficiencies.}
}
\value{
A data frame of the final energy to useful exergy multipliers
}
\description{
Calculates the final energy to useful exergy multipliers
}
