% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiencies.R
\name{calc_fu_Y_EIOU_agg_efficiencies}
\alias{calc_fu_Y_EIOU_agg_efficiencies}
\title{Calculates the aggregated FU efficiency of products when used in EIOU, Y, or economy-wide}
\usage{
calc_fu_Y_EIOU_agg_efficiencies(
  C_mats_agg,
  eta_fu_vecs,
  phi_vecs,
  non_energy_use_machine = "Non-energy consumption -> NEU",
  eta.fu = "eta.fu",
  C_EIOU_agg = "C_EIOU_agg",
  C_Y_agg = "C_Y_agg",
  C_EIOU_Y_agg = "C_EIOU_Y_agg",
  C_EIOU_agg_excl_NEU = "C_EIOU_agg_excl_NEU",
  C_Y_agg_excl_NEU = "C_Y_agg_excl_NEU",
  C_EIOU_Y_agg_excl_NEU = "C_EIOU_Y_agg_excl_NEU",
  eta_p_eiou = "eta_p_eiou",
  eta_p_y = "eta_p_y",
  eta_p_eiou_y = "eta_p_eiou_y",
  phi = "phi",
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  product = IEATools::iea_cols$product
)
}
\arguments{
\item{C_mats_agg}{A data frame containing the aggregated C matrices.}

\item{eta_fu_vecs}{A data frame containing the efficiency vectors.}

\item{phi_vecs}{A data frame containing the phi values.}

\item{non_energy_use_machine}{The character string of the non-energy use machine that needs to be excluded
for efficiencies excluding non-energy uses.}

\item{eta.fu}{The name of the column containing the machine efficiencies in the eta_fu_vecs data frame.}

\item{C_EIOU_agg}{The name of the column containing the C_EIOU aggregated matrix in the output data frame.}

\item{C_Y_agg}{The name of the column containing the C_Y aggregated matrix in the output data frame.}

\item{C_EIOU_Y_agg}{The name of the column containing the C_EIOU_Y aggregated matrix in the output data frame.}

\item{C_EIOU_agg_excl_NEU}{The name of a temporary column containing the C_EIOU aggregated matrix excluding non-energy uses.}

\item{C_Y_agg_excl_NEU}{The name of a temporary column containing the C_Y aggregated matrix excluding non-energy uses.}

\item{C_EIOU_Y_agg_excl_NEU}{The name of a temporary column containing the C_EIOU_Y aggregated matrix excluding non-energy uses.}

\item{eta_p_eiou}{The name of the column containing the efficiency of each product when used in EIOU.}

\item{eta_p_y}{The name of the column containing the efficiency of each product when used in final demand.}

\item{eta_p_eiou_y}{The name of the column containing the efficiency of each product when used in either EIOU or final demand.}

\item{phi}{The name of the column containing the phi values.}

\item{country, method, energy_type, last_stage, year, product}{See \code{IEATools::iea_cols}.}
}
\value{
A data frame containing the product efficiencies when used in EIOU, Y, or economy-wide (Y and EIOU).
}
\description{
Calculates the aggregated FU efficiency of products when used in EIOU, Y, or economy-wide
}
