% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_to_exergy.R
\name{calc_phi_vecs_mw}
\alias{calc_phi_vecs_mw}
\title{Create a data frame of phi vectors for muscle work}
\usage{
calc_phi_vecs_mw(
  psut_energy_mw,
  phi_vec_mw,
  countries,
  country = MWTools::mw_cols$country,
  year = MWTools::mw_cols$year,
  phi = "phi"
)
}
\arguments{
\item{psut_energy_mw}{A PSUT data frame containing \code{country} and \code{year} columns.}

\item{phi_vec_mw}{A single vector of muscle work phi values. See \code{MWTools::phi_vec_mw()}.}

\item{countries}{The countries to be analyzed. Internally, \code{psut_energy_mw} is filtered for \code{countries}.}

\item{country, year}{Column names. See \code{MWTools::mw_cols}.}

\item{phi}{The name of the phi column. Default is "phi".}
}
\value{
A data frame of muscle work phi vectors, suitable for \code{move_to_exergy()}.
}
\description{
The \code{phi_vecs} argument of \code{move_to_exergy()} is a data frame with
"Country", "Year", and "phi" columns,
where the "phi" column contains phi vectors of the type created by
\code{MWTools::phi_vec_mw()}.
This function creates the required data frame from its parts.
}
