% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canonical_countries}
\alias{canonical_countries}
\title{Canonical countries}
\format{
A string list with 154 entries.
}
\usage{
canonical_countries
}
\description{
Countries and country groups which, if aggregated, cover the entire world
(WRLD) without double counting.
}
\examples{
canonical_countries
}
\keyword{datasets}
