% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{combine_countries_exemplars}
\alias{combine_countries_exemplars}
\title{Combine countries and additional exemplars}
\usage{
combine_countries_exemplars(couns, exempls)
}
\arguments{
\item{couns}{Countries of interest}

\item{exempls}{Additional exemplars}
}
\value{
A unique combination of countries and additional exemplars
}
\description{
Combine countries and additional exemplars
}
