% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{double_counted_countries}
\alias{double_counted_countries}
\title{Double-counted countries}
\format{
A string list with 17 entries.
}
\usage{
double_counted_countries
}
\description{
Selected individual countries and selected country groups which, if aggregated,
would result in double counting.
}
\examples{
double_counted_countries
}
\keyword{datasets}
