% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exemplar_names}
\alias{exemplar_names}
\title{Exemplar table names}
\format{
A string list with 16 entries.
\describe{
\item{exemplar_tab_name}{The string name of the tab in the Excel file containing the exemplar table.}
\item{prev_names}{The name of a column of previous names used for the country.}
\item{exemplars}{The name of a column of exemplar countries.}
\item{exemplar_country}{The name of an exemplar country column.}
\item{exemplar_countries}{The name of an exemplar countries column.}
\item{exemplar_tables}{The name of a column containing exemplar tables.}
\item{iea_data}{The name of a column containing IEA extended energy balance data.}
\item{alloc_data}{The name of a column containing final-to-useful allocation data.}
\item{incomplete_alloc_table}{The name of a column containing incomplete final-to-useful allocation tables.}
\item{complete_alloc_table}{The name of a column containing completed final-to-useful allocation tables.}
\item{incomplete_eta_table}{The name of a column containing incomplete final-to-useful efficiency tables.}
\item{complete_eta_table}{The name of a column containing completed final-to-useful efficiency tables.}
\item{region_code}{The name of the region code column.}
\item{country_name}{The name of the column containing the long name of a country.}
\item{agg_code_col}{The metadata column "Agg.Code", representing the country, or country group code for individual country level data to be aggregated in to.}
\item{world}{The name of the world region.}
}
}
\usage{
exemplar_names
}
\description{
A string list containing named names of columns and tabs for exemplar tables.
Items in the list provide default values for column name function arguments
throughout the \code{PFUPipeline} package.
}
\examples{
exemplar_names
}
\keyword{datasets}
