% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_functions.R
\name{expand_phi_vecs}
\alias{expand_phi_vecs}
\title{Prepare phi vectors for writing as a data frame}
\usage{
expand_phi_vecs(
  phi_vecs,
  machine = "Machine",
  phi = "phi",
  matnames = "matnames",
  matvals = "matvals",
  rownames = "rownames",
  colnames = "colnames",
  rowtypes = "rowtypes",
  coltypes = "coltypes"
)
}
\arguments{
\item{phi_vecs}{The incoming Phivecs target.}

\item{machine, phi, matnames, matvals, rownames, colnames, rowtypes, coltypes}{Columns names in \code{phi_vecs}.}
}
\value{
A data frame with columns Country, Year, Machine, and phi.
}
\description{
The Phivecs target is a \code{matsindf} data frame
that contains a column of phi vectors.
But another common data format is a simple data frame.
This function converts the Phivecs matsindf data frame to
a data table.
}
