% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_to_exergy.R
\name{extend_details_matrices_to_exergy}
\alias{extend_details_matrices_to_exergy}
\title{Move the useful details energy matrices to exergy}
\usage{
extend_details_matrices_to_exergy(
  fu_details_mats,
  phi_vecs,
  countries,
  country_colname = IEATools::iea_cols$country,
  year_colname = IEATools::iea_cols$year
)
}
\arguments{
\item{fu_details_mats}{}

\item{phi_vecs}{}

\item{countries}{The countries for which this function should be applied.}

\item{country_colname}{}

\item{year_colname}{}

\item{exergy_type}{}

\item{energy_type_colname}{}

\item{phi_colname}{}
}
\value{
A version of \code{fu_details_mats} with matrices containing exergy at the useful stage.
}
\description{
A function to extend a data frame of energy fu details matrices to exergy.
}
\details{
The final-to-useful details matrices contain information in rows and columns
about the transition from final to useful energy and exergy,
including
\itemize{
\item the final energy product,
\item the final demand sector,
\item the final-to-useful machine, and
\item the the useful energy product.
}

Entries in the details matrices are useful energy amounts.
Information is encoded in
row and column labels of the details matrix:
\itemize{
\item Row names use \code{RCLabels::arrow_notation} with
\itemize{
\item a prefix that identifies the final energy product and
\item a suffix that identifies the sector in which the final-to-useful
transformation occurs.
\item Example: "Aviation gasoline -> Domestic aviation".
}
\item Column names use \code{RCLabels::from_notation} with
\itemize{
\item a prefix that identifies the useful energy product and
\item a suffix that identifies the final-to-useful machine.
\item Example: "HPL \link{from Electric pumps}".
}
}

This function enables mapping over countries.
}
