% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_iea_mw.R
\name{filter_mw_to_iea_years}
\alias{filter_mw_to_iea_years}
\title{Filter muscle work data to only those years contained in IEA data}
\usage{
filter_mw_to_iea_years(
  .psut_mw,
  .psut_iea,
  countries,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage
)
}
\arguments{
\item{.psut_mw}{The incoming muscle work data frame to be filtered.}

\item{.psut_iea}{The incoming IEA data frame from which years are obtained.}

\item{countries}{The countries for which the filtering should be done.}

\item{country, year, method, energy_type, last_stage}{The columns in \code{.psut_mw} and \code{.iea_mw}
to be used for filtering.}
}
\value{
A version of \code{.psut_mw} that contains only those
countries, years, methods, energy types, and last stages
also contained in \code{.psut_iea}.
}
\description{
This function compares the muscle work and IEA data frames
on the \code{country}, \code{year}, \code{method}, \code{energy_type}, and \code{last_stage} columns
and keeps only those rows in \code{.psut_mw} that match \code{.psut_iea}.
}
