% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socioecon_functions.R
\name{get_all_pwt_data}
\alias{get_all_pwt_data}
\title{Get all pwt10 data for a set of countries}
\usage{
get_all_pwt_data(
  countries,
  isocode_colname = PFUDatabase::socioecon_cols$isocode_colname
)
}
\arguments{
\item{countries}{A string of iso 3-letter country codes}

\item{isocode_colname}{See \code{PFUDatabase::socioecon_cols}.}
}
\value{
A data frame containing socioeconomic data from pwt10 for a set of countries.
}
\description{
Using the pwt10 package this function creates a data frame containing all the
data Penn World Tables 10 (pwt10) for a set of countries as specified in a string
of 3-letter ISO country codes named \code{countries}.
Note that some data is not available for some countries (mostly non-OECD).
}
\examples{
countries <- c("GBR")
all_pwt_data <- get_all_pwt_data(countries = countries)
}
