% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_fd_sectors}
\alias{get_fd_sectors}
\title{Retrieve a list of final demand sectors}
\usage{
get_fd_sectors()
}
\value{
A list of final demand sectors from \code{IEATools::fd_sectors}.
}
\description{
Retrieve a list of final demand sectors for calculation of the total final consumption
of final, useful, or services energy in gross or net terms.
}
\examples{
fd_sectors <- get_fd_sectors()
}
