% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socioecon_functions.R
\name{get_L_K_GDP_data}
\alias{get_L_K_GDP_data}
\title{Create a dataframe containing capital (K), labor (L), and GDP data}
\usage{
get_L_K_GDP_data(
  pwt10_data,
  country_colname = IEATools::iea_cols$country,
  Year_colname = IEATools::iea_cols$year,
  year_colname = PFUDatabase::socioecon_cols$year_colname,
  isocode_colname = PFUDatabase::socioecon_cols$isocode_colname,
  rgdpe_colname = PFUDatabase::socioecon_cols$rgdpe_colname,
  rgdpo_colname = PFUDatabase::socioecon_cols$rgdpo_colname,
  rgdpna_colname = PFUDatabase::socioecon_cols$rgdpna_colname,
  emp_colname = PFUDatabase::socioecon_cols$emp_colname,
  avh_colname = PFUDatabase::socioecon_cols$avh_colname,
  hc_colname = PFUDatabase::socioecon_cols$hc_colname,
  rnna_colname = PFUDatabase::socioecon_cols$rnna_colname,
  rkna_colname = PFUDatabase::socioecon_cols$rkna_colname,
  K_colname = PFUDatabase::socioecon_cols$K_colname,
  Kserv_colname = PFUDatabase::socioecon_cols$Kserv_colname,
  L_colname = PFUDatabase::socioecon_cols$L_colname,
  Ladj_colname = PFUDatabase::socioecon_cols$Ladj_colname
)
}
\arguments{
\item{pwt10_data}{A data frame containing all pwt10 data for at least one country,
usually obtained from \code{get_all_pwt_data()}.}

\item{country_colname, Year_colname}{See \code{IEATools::iea_cols}.}

\item{isocode_colname, year_colname, rgdpe_colname, rgdpo_colname, rgdpna_colname, emp_colname, avh_colname, hc_colname, rnna_colname, rkna_colname, K_colname, Kserv_colname, L_colname, Ladj_colname}{See \code{PFUDatabase::socioecon_cols}.}
}
\value{
A data frame containing three GDP metrics, Labor, Adjusted Labor,
Capital, and Capital services.
}
\description{
This function selects the following columns from a pwt10 data frame produced
by calling \code{pwt10::pwt10.0},
with descriptions from pwt10 documentation:
isocode: 3-letter isocode
year: Year
rgdpe: Expenditure-side real GDP at chained PPPs (in million 2017 USD).
rgdpo: Output-side real GDP at chained PPPs (in million 2017 USD).
rgdpna: Real GDP at constant 2017 national prices (in million 2017 USD)
emp: Number of persons engaged (in millions)
avh: Average annual hours worked by persons engaged.
hc: Human capital index, based on years of schooling and returns to education;
see Human capital in PWT9.
rnna: Capital stock at constant 2017 national prices (in million 2017 USD).
rkna: Capital services at constant 2017 national prices (2017 = 1).
}
\details{
The metrics  L, the total number of hours worked in a given year
and Ladj, the number of hours worked adjusted by the human capital index are
also calculated and added as columns, with avh, hc, and emp being removed after use.

Note that some data is not available for some countries (mostly non-OECD),
some of the calculated metrics i.e. Adjusted Labor (L.adj) is also absent.
}
\examples{
countries <- c("GBR")
L_K_GDP_data <- get_all_pwt_data(countries = countries) \%>\%
                  get_L_K_GDP_data()
}
