% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_p_industry_prefixes}
\alias{get_p_industry_prefixes}
\title{Retrieve primary industry prefixes}
\usage{
get_p_industry_prefixes()
}
\value{
A list of primary industry prefixes from \code{IEATools::prim_agg_flows}.
}
\description{
Retrieve primary industry prefixes for use by \code{Recca::find_p_industry_names}.
Contains "Resources", "Imports", and "Stock changes".
}
\examples{
p_industry_prefixes <- get_p_industry_prefixes()
}
