% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ieamw_cols}
\alias{ieamw_cols}
\title{Constants for data frames containing IEA and MW data frames}
\format{
A string list with 4 entries.
\describe{
\item{ieamw}{The name of the column containing metadata on ECC sources. "IEAMW"}
\item{iea}{A string identifying that ECC data are from the IEA exclusively. "IEA"}
\item{mw}{A string identifying that ECC data are for muscle work (MW) exclusively. "MW"}
\item{both}{A string identifying that ECC data include both IEA and muscle work. "Both"}
}
}
\usage{
ieamw_cols
}
\description{
A string list containing both the column name and column values
for energy conversion chains (ECCs) in PSUT format.
}
\examples{
ieamw_cols
}
\keyword{datasets}
