% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_functions.R
\name{load_amw_pfu_data}
\alias{load_amw_pfu_data}
\title{Load animal muscle work data}
\usage{
load_amw_pfu_data(fao_data_path, mw_concordance_path, amw_analysis_data_path)
}
\arguments{
\item{fao_data_path}{The path to FAO data.}

\item{mw_concordance_path}{The path to the muscle work concordance.}

\item{amw_analysis_data_path}{The path the animal muscle work data.}
}
\value{
A data frame of animal muscle work data.
}
\description{
This function loads animal muscle work data and
renames the sectors according to the default arguments to \code{rename_mw_sectors()}.
}
