% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exemplar_functions.R
\name{load_exemplar_table}
\alias{load_exemplar_table}
\title{Read an exemplar table from an Excel file}
\usage{
load_exemplar_table(
  exemplar_table_path = sample_exemplar_table_path(),
  countries = NULL,
  years = NULL,
  exemplar_table_tab_name = PFUPipeline::exemplar_names$exemplar_tab_name,
  prev_names = PFUPipeline::exemplar_names$prev_names,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year
)
}
\arguments{
\item{exemplar_table_path}{The path to the Excel file containing an exemplar table.
Default is the value of \code{sample_exemplar_table_path()}.}

\item{countries}{The countries for which exemplars are desired. If \code{NULL}, the default,
all countries in the file at \code{exemplar_table_path} are returned.}

\item{years}{The years for which you want the exemplar table, an integer.
Default is \code{NULL}, meaning that all years from the file at \code{exemplar_table_path} are included.}

\item{exemplar_table_tab_name, prev_names}{See \code{PFUPipeline::exemplar_names}.}

\item{year, country}{See \code{IEATools::iea_cols}.}
}
\value{
an exemplar table
}
\description{
An exemplar table has a column of country names
and additional metadata,
including the exemplar country and any name changes through time.
}
\details{
This function gathers (actually, \code{tidyr::pivot_longer()}) years into a Year column.
}
\examples{
load_exemplar_table()
}
