% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_functions.R
\name{load_hmw_pfu_data}
\alias{load_hmw_pfu_data}
\title{Load human muscle work data}
\usage{
load_hmw_pfu_data(ilo_data_path, mw_concordance_path, hmw_analysis_data_path)
}
\arguments{
\item{ilo_data_path}{The path to ILO data.}

\item{mw_concordance_path}{The path to the muscle work concordance.}

\item{hmw_analysis_data_path}{The path the human muscle work data.}
}
\value{
A data frame of human muscle work data.
}
\description{
This function loads human muscle work data and
renames the sectors according to the default arguments to \code{rename_mw_sectors()}.
}
