% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{machine_constants}
\alias{machine_constants}
\title{Constants for machine data files}
\format{
A string list with 1 entries.
\describe{
\item{efficiency_tab_name}{The name of the machine data efficiency tab. Default is "FIN_ETA".}
}
}
\usage{
machine_constants
}
\description{
A string list containing values for features of machine data files.
}
\examples{
machine_constants
}
\keyword{datasets}
