% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{make_balanced}
\alias{make_balanced}
\title{Balance IEA data}
\usage{
make_balanced(
  .iea_data,
  max_fix = 6,
  countries,
  country = IEATools::iea_cols$country,
  grp_vars = c(country, IEATools::iea_cols$method, IEATools::iea_cols$energy_type,
    IEATools::iea_cols$last_stage, IEATools::iea_cols$year, IEATools::iea_cols$product)
)
}
\arguments{
\item{.iea_data}{A tidy IEA data frame.}

\item{max_fix}{The maximum allowable energy imbalance to fix.
Default is \code{3}.}

\item{countries}{The countries that should be balanced.}

\item{country}{See \code{IEATools::iea_cols}}

\item{grp_vars}{the groups that should be checked.
Default is
\code{c(country, IEATools::iea_cols$method, IEATools::iea_cols$energy_type, IEATools::iea_cols$last_stage, IEATools::iea_cols$product)}.}
}
\value{
A data frame of balanced IEA data.
}
\description{
Balances the IEA data in a way that is amenable to drake subtargets.
Internally, this function uses \code{IEATools::fix_tidy_iea_df_balances()}.
Grouping is done internal to this function using the value of \code{grp_vars}.
}
\examples{
IEATools::sample_iea_data_path() \%>\%
  IEATools::load_tidy_iea_df() \%>\%
  make_balanced(countries = c("GHA", "ZAF")) \%>\%
  is_balanced(countries = c("GHA", "ZAF"))
}
