% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{make_iea_psut}
\alias{make_iea_psut}
\title{Convert to PSUT matrices}
\usage{
make_iea_psut(
  SpecifiedIEAData,
  countries,
  matrix_class = c("matrix", "Matrix"),
  country = IEATools::iea_cols$country
)
}
\arguments{
\item{SpecifiedIEAData}{A data frame that has already been specified via \code{specify()}.}

\item{countries}{The countries you want to convert to PSUT matrices.}

\item{matrix_class}{The type of matrix to be created.
One of "matrix" (the base class) or
"Matrix" (for sparse matrices).}

\item{country}{See \code{IEATools::iea_cols}.}
}
\value{
A \code{matsindf}-style data frame.
}
\description{
Converts tidy IEA data to PSUT matrices in a way that is amenable to drake subtargets.
Internally, \code{IEATools::prep_psut()} does the conversion to matrices.
}
\examples{
IEATools::sample_iea_data_path() \%>\%
  IEATools::load_tidy_iea_df() \%>\%
  make_balanced(countries = c("GHA", "ZAF")) \%>\%
  specify(countries = c("GHA", "ZAF")) \%>\%
  make_iea_psut(countries = c("GHA", "ZAF"))
}
