% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_functions.R
\name{make_mw_psut}
\alias{make_mw_psut}
\title{Calculate PSUT data frames from muscle work PFU data frames}
\usage{
make_mw_psut(
  .hmw_df,
  .amw_df,
  countries,
  years,
  country = MWTools::mw_cols$country,
  year = MWTools::mw_cols$year
)
}
\arguments{
\item{.hmw_df, .amw_df}{Data frames of primary-final-useful muscle work data.}

\item{countries}{The countries to be analyzed.}

\item{years}{The years to be analyzed.}

\item{country}{The name of the country column in \code{.hmw_df} and \code{.amw_df}. Default is \code{MWTools::mw_cols$country}.}

\item{year}{The name of the year column in \code{.hmw_df} and \code{.amw_df}.Default is \code{MWTools::mw_cols$year}.}
}
\value{
A data frame of PSUT matrices for a muscle work energy conversion chain.
}
\description{
This function filters to \code{countries} and creates PSUT matrices
from primary-final-useful muscle work data frames.
}
