% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_to_useful.R
\name{move_to_useful_with_details}
\alias{move_to_useful_with_details}
\title{Move the last stage of the energy conversion chain from final stage to useful stage with details}
\usage{
move_to_useful_with_details(
  psut_final,
  C_mats,
  eta_phi_vecs,
  countries,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  C_Y = IEATools::template_cols$C_Y,
  C_eiou = IEATools::template_cols$C_eiou
)
}
\arguments{
\item{psut_final}{A data frame with rows that describe the energy conversion chain with final energy as the last stage.}

\item{C_mats}{A data frame with allocation matrices, probably the Cmats target.}

\item{eta_phi_vecs}{A data frame with final-to-useful efficiency and exergy-to-energy ratio vectors.}

\item{countries}{The countries to be analyzed.}

\item{country, year}{See IEATools::iea_cols.}

\item{C_Y, C_eiou}{See IEATools::template_cols.}
}
\value{
A data frame with energy conversion chain matrices with last stage as useful energy.
}
\description{
Extends the energy conversion chain from a final energy last stage to useful energy last stage.
Details about the conversion from final to useful are retained via matrices
\strong{Y_fu_details} and \strong{U_EIOU_fu_details}.
The last-stage-useful energy conversion chain PSUT matrices are bound as rows at the bottom
of the \code{psut_final} data frame,
albeit with "Useful" instead of "Final" in the \code{Last.stage} column.
}
