% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phi_sources}
\alias{phi_sources}
\title{Sources for phi values}
\format{
A string list with 3 entries.
\describe{
\item{eta_fu_tables}{Tables of final-to-useful efficiency values.}
\item{temperature_data}{Country-average yearly temperature data.}
\item{phi_constants}{Tables of constant phi values.}
}
}
\usage{
phi_sources
}
\description{
A string list containing named sources of phi (exergy-to-energy ratio) values.
}
\examples{
phi_sources
}
\keyword{datasets}
