% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{phi_u_graph}
\alias{phi_u_graph}
\title{Generate an phi_u graph}
\usage{
phi_u_graph(
  .df,
  countries,
  country = IEATools::iea_cols$country,
  quantity = IEATools::template_cols$quantity,
  year = IEATools::iea_cols$year,
  .values = IEATools::template_cols$.values,
  machine = IEATools::template_cols$machine,
  eu_product = IEATools::template_cols$eu_product,
  machine_eu_product = paste0(machine, "_", eu_product)
)
}
\arguments{
\item{.df}{A data frame comprised of completed exergy-to-energy ratio values}

\item{countries}{The countries for which efficiency plots are to be created.}

\item{country, year}{See \code{IEATools::iea_cols}.}

\item{.values, machine, quantity, eu_product}{See \code{IEATools::template_cols}.}

\item{machine_eu_product}{The name of a combined \code{machine} and \code{eu_product} column.}
}
\value{
A \code{ggplot2} graph object
}
\description{
Creates an phi_u graph from a completed eta_fu table,
created by the \code{assemble_eta_fu_tables()} function.
}
\details{
This function is called repeatedly from \code{phi_u_plots_df()}.

\code{machine}, and \code{destination} form the title of the graph.
}
\examples{
library(ggplot2)
# Make a simple data frame with the expected structure.
tibble::tribble(~Year, ~.values, ~Machine, ~Eu.product,
                1967, 0.5, "Cars", "MD",
                1967, 0.5, "Industry static engines", "MD",
                2020, 0.8, "Cars", "MD",
                2020, 0.2, "Industry static engines", "MD") \%>\%
  alloc_graph(country = "Example", ef_product = "Petrol", destination = "Transport")
}
