% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exiobase.R
\name{read_country_concordance_table}
\alias{read_country_concordance_table}
\title{Reads the country concordance table, and keeps countries in the database}
\usage{
read_country_concordance_table(country_concordance_table_file, countries)
}
\arguments{
\item{country_concordance_table_file}{The name of the file containing the counctry concordance table}

\item{countries}{Contains the list of countries included in the run of the database}
}
\value{
A data frame containing the country concordance table
}
\description{
Reads the country concordance table, and keeps countries in the database
}
