% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exiobase.R
\name{read_list_exiobase_energy_flows}
\alias{read_list_exiobase_energy_flows}
\title{Reads the list of Exiobase energy flows}
\usage{
read_list_exiobase_energy_flows(path_to_list_exiobase_energy_flows)
}
\arguments{
\item{path_to_list_exiobase_energy_flows}{Contains the path to the folder containing the list of Exiobase energy flows}
}
\value{
A data frame containing the list of Exiobase energy flows, and a boolean column stating whether the flow is a final energy flow or not
}
\description{
Reads the list of Exiobase energy flows
}
