% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storing.R
\name{release_target}
\alias{release_target}
\title{Save a target to a pinboard.}
\usage{
release_target(
  pipeline_releases_folder,
  targ,
  targ_name,
  type = "rds",
  release = FALSE
)
}
\arguments{
\item{pipeline_releases_folder}{The folder that contains the pinboard for releases from the pipeline.}

\item{targ}{The target R object to be saved to the pinboard.}

\item{targ_name}{The name of the target object. \code{targ_name} is the key to retrieving \code{targ}.}

\item{type}{The type of the target, routed to \code{pins::pin_write()}. Default is "rds".}

\item{release}{A boolean telling whether to do a release.
Default is \code{FALSE}.}
}
\value{
If \code{release} is \code{TRUE},
the fully-qualified path name of the \code{targ} file in the pinboard.
If \code{release} is \code{FALSE}, the string "Release not requested."
}
\description{
Releases (\code{release = TRUE})
or not (\code{release = FALSE})
a new version of the target target
using the \code{pins} package.
}
\details{
Released versions of the target can be obtained
as shown in examples.
}
\examples{
\dontrun{
# Establish the pinboard
pinboard <- pins::board_folder("~/Dropbox/Fellowship 1960-2015 PFU database/PipelineReleases/")
# Get information about the `PSUT` target in the pinboard
pinboard \%>\%
  pins::pin_meta(name = "psut")
# Find versions of the `PSUT` target
pinboard \%>\%
  pins::pin_versions(name = "psut")
# Get the latest copy of the `PSUT` target.
my_psut <- pinboard \%>\%
  pins::pin_read(name = "psut")
# Retrieve a previous version of the `PSUT` target.
my_old_psut <- pinboard \%>\%
  pins::pin_read(name = "psut", version = "20220218T023112Z-1d9e1")}
}
