% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_to_useful.R
\name{remove_cols_from_PSUTUsefulIEAWithDetails}
\alias{remove_cols_from_PSUTUsefulIEAWithDetails}
\title{Remove columns from the PSUTUsefulIEAWithDetails target}
\usage{
remove_cols_from_PSUTUsefulIEAWithDetails(
  psut_useful_iea_with_details,
  cols_to_remove,
  phi_vecs,
  remove_final = FALSE,
  countries,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  last_stage = IEATools::iea_cols$last_stage,
  final = IEATools::all_stages$final
)
}
\arguments{
\item{psut_useful_iea_with_details}{The target from which columns should be removed.}

\item{cols_to_remove}{A string vector of columns names to be removed.}

\item{remove_final}{A boolean that tells whether to remove Last.stage == "Final" rows.
Default is \code{FALSE}.}

\item{countries}{The countries for which this function should be applied.}

\item{country}{The name of the \code{Country} column in \code{psut_useful_iea_with_details} and \code{phi_vecs}.
Default is \code{IEATools::iea_cols$country}.}

\item{year}{The name of the \code{Year} column in  \code{psut_useful_iea_with_details} and \code{phi_vecs}.
Default is \code{IEATools::iea_cols$year}.}

\item{last_stage}{The name of the "Last.stage" column.
Default is \code{IEATools::iea_cols$last_stage}.}

\item{final}{The string defining the final stage.
Default is \code{IEATools::all_stages$final}.}
}
\value{
A version of \code{psut_useful_iea_with_details} with \code{cols_to_remove} removed.
}
\description{
A simple wrapper function to assist with removing
unneeded columns from the \code{PSUTUsefulIEAWithDetails} target.
}
\details{
This function enables mapping over countries.
}
