% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_non_energy_use.R
\name{remove_non_energy_use}
\alias{remove_non_energy_use}
\title{Remove Non-energy use from energy conversion chains}
\usage{
remove_non_energy_use(.psut_mats)
}
\arguments{
\item{.psut_mats}{A data frame of PSUT matrices.}
}
\value{
A version of \code{.psut_mats} with Non-energy use removed.
}
\description{
It is helpful to remove Non-energy use from ECCs.
This function calls \code{Recca::remove_neu()}.
}
