% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exemplar_functions.R
\name{sample_exemplar_table_path}
\alias{sample_exemplar_table_path}
\title{Gives a file path to a sample exemplar table}
\usage{
sample_exemplar_table_path()
}
\value{
A path to a sample exemplar table bundled with this package.
}
\description{
Gives a file path to a sample exemplar table
}
\examples{
sample_exemplar_table_path()
}
