% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_to_useful.R
\name{sep_eta_fu_phi_u}
\alias{sep_eta_fu_phi_u}
\title{Choose eta.fu or phi.u columns from a data frame of eta.fu and phi.u vectors.}
\usage{
sep_eta_fu_phi_u(
  eta_fu_phi_u_vecs,
  keep = c(IEATools::template_cols$eta_fu, IEATools::template_cols$phi_u),
  countries,
  country = IEATools::iea_cols$country
)
}
\arguments{
\item{eta_fu_phi_u_vecs}{A data frame containing metadata columns and columns for
eta_fu (final to useful efficiency) and
phi_u (exergy-to-energy efficiency ratios).}

\item{keep}{Tells which column to keep, eta_fu or phi_u.
Must be one of \code{IEATools::template_cols$eta_fu} or \code{IEATools::template_cols$phi_u}.}

\item{countries}{The countries to be analyzed.}

\item{country}{See IEATools::iea_cols.}
}
\value{
A data frame of metadata and either an eta_fu column or a phi_u column
}
\description{
Choose eta.fu or phi.u columns from a data frame of eta.fu and phi.u vectors.
}
