% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{socioecon_cols}
\alias{socioecon_cols}
\title{Constants for socioeconomic data files}
\format{
A string list with 14 entries.
\describe{
\item{isocode_colname}{The name of the column containing ISO country codes, "isocode".}
\item{year_colname}{The name of the column containing years, "Year".}
\item{rgdpe_colname}{The name of the column containing rgdpe values, "rgdpe".}
\item{rgdpo_colname}{The name of the column containing rgdpo values, "rgdpo".}
\item{rgdpna_colname}{The name of the column containing rgdpna values, "rgdpna".}
\item{emp_colname}{The name of the column containing employment values, "emp".}
\item{avh_colname}{The name of the column containing avh values, "avh".}
\item{hc_colname}{The name of the column containing hc values, "hc".}
\item{rnna_colname}{The name of the column containing rnna values, "rnna".}
\item{rkna_colname}{The name of the column containing rkna values, "rkna".}
\item{K_colname}{The name of the column containing capital values, "K".}
\item{Kserv_colname}{The name of the column containing capital services values, "Kserv".}
\item{L_colname}{The name of the column containing labor values, "L".}
\item{Ladj_colname}{The name of the column containing quality-adjusted labor values, "Ladj".}
}
}
\usage{
socioecon_cols
}
\description{
A string list containing values for features of socioeconomic data files.
}
\examples{
socioecon_cols
}
\keyword{datasets}
