% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_functions.R
\name{verify_mw_energy_balance}
\alias{verify_mw_energy_balance}
\title{Verify energy balance in muscle work PSUT matrices}
\usage{
verify_mw_energy_balance(
  .psut_df,
  countries,
  country = MWTools::mw_cols$country
)
}
\arguments{
\item{.psut_df}{A data frame of muscle work PSUT matrices.}

\item{countries}{The countries to be analyzed.}

\item{country}{The name of the country column in \code{.psut_df}.}
}
\value{
A data frame with new boolean column ".balanced" that tells
whether the matrices are balanced.
}
\description{
After constructing the muscle work PSUT matrices,
energy balance should be verified.
Internally, this function uses \code{Recca::verify_SUT_energy_balance()}
to ensure that everything is balanced.
}
\details{
If \code{.psut_df} has zero rows,
\code{TRUE} is returned, enabling the pipeline to continue,
even if there are some years where there is no muscle work data available.
}
