% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_iea_mw.R
\name{add_iea_mw_psut}
\alias{add_iea_mw_psut}
\title{Sum IEA and muscle work ECC matrices}
\usage{
add_iea_mw_psut(
  .iea_psut,
  .mw_psut,
  countries,
  R = IEATools::psut_cols$R,
  U = IEATools::psut_cols$U,
  V = IEATools::psut_cols$V,
  Y = IEATools::psut_cols$Y,
  U_feed = IEATools::psut_cols$U_feed,
  U_eiou = IEATools::psut_cols$U_eiou,
  s_units = IEATools::psut_cols$s_units,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  r_eiou = IEATools::psut_cols$r_eiou
)
}
\arguments{
\item{.iea_psut}{An IEA PSUT data frame.}

\item{.mw_psut}{A muscle work PSUT data frame.}

\item{countries}{The countries to be analyzed.}

\item{R}{The name of the column of \code{R} matrices. Default is \code{IEATools::psut_cols$R}.}

\item{U}{The name of the column of \code{U} matrices. Default is \code{IEATools::psut_cols$U}.}

\item{V}{The name of the column of \code{V} matrices. Default is \code{IEATools::psut_cols$V}.}

\item{Y}{The name of the column of \code{Y} matrices. Default is \code{IEATools::psut_cols$Y}.}

\item{U_feed}{The name of the column of \code{U_feed} matrices. Default is \code{IEATools::psut_cols$U_feed}.}

\item{U_eiou}{The name of the column of \code{U_eiou} matrices. Default is \code{IEATools::psut_cols$U_eiou}.}

\item{s_units}{The name of the column of \code{s_units} matrices. Default is \code{IEATools::psut_cols$s_units}.}

\item{country}{The name of the country column. Default is \code{IEATools$iea_cols$country}.}

\item{year}{The name of the year column. Default is \code{IEATools$iea_cols$year}.}

\item{method}{The name of the method column. Default is \code{IEATools$iea_cols$method}.}

\item{energy_type}{The name of the energy type column. Default is \code{IEATools$iea_cols$energy_type}.}

\item{last_stage}{The name of the last_stage column. Default is \code{IEATools$iea_cols$last_stage}.}

\item{r_eiou}{The name of the r_eiou column. Default is \code{IEATools$iea_cols$r_eiou}.}
}
\value{
A data frame of summed matrices.
}
\description{
To create a combined energy conversion chain (ECC)
containing both IEA and muscle work data,
PSUT matrices for each ECC are summed.
This function sums \code{R}, \code{U}, \code{V}, \code{Y}, \code{U_feed}, and \code{U_eiou} matrices directly.
It also re-calculates the \code{r_eiou} matrix.
}
\details{
If either of \code{iea_psut} or \code{.mw_psut} are \code{NULL},
the other is returned.
}
