% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_iea_mw.R
\name{build_psut_dataframe}
\alias{build_psut_dataframe}
\title{Build the final PSUT data frame}
\usage{
build_psut_dataframe(
  psutiea = NULL,
  psutmw = NULL,
  psutieamw = NULL,
  country_colname = IEATools::iea_cols$country,
  method_colname = IEATools::iea_cols$method,
  energy_type_colname = IEATools::iea_cols$energy_type,
  last_stage_colname = IEATools::iea_cols$last_stage,
  year_colname = IEATools::iea_cols$year,
  ieamw_colname = PFUDatabase::ieamw_cols$ieamw,
  R_colname = IEATools::psut_cols$R,
  U_colname = IEATools::psut_cols$U,
  U_feed_colname = IEATools::psut_cols$U_feed,
  U_eiou_colname = IEATools::psut_cols$U_eiou,
  r_eiou_colname = IEATools::psut_cols$r_eiou,
  V_colname = IEATools::psut_cols$V,
  Y_colname = IEATools::psut_cols$Y,
  S_units_colname = IEATools::psut_cols$s_units,
  iea = PFUDatabase::ieamw_cols$iea,
  mw = PFUDatabase::ieamw_cols$mw,
  both = PFUDatabase::ieamw_cols$both
)
}
\arguments{
\item{psutiea}{A PSUT data frame of IEA data. Default is \code{NULL.}}

\item{psutmw}{A PSUT data frame of muscle work data. Default is \code{NULL.}}

\item{psutieamw}{A PSUT data frame of combined IEA and MW data. Default is \code{NULL.}}

\item{country_colname, method_colname, energy_type_colname, last_stage_colname, year_colname}{Column names.
See \code{IEATools::iea_cols} for defaults.}

\item{ieamw_colname}{The name of the column that identifies whether data are for the IEA,
muscle work (MW) or both.
Default is \code{PFUDatabase::ieamw_cols$ieamw}.}

\item{R_colname, U_colname, U_feed_colname, U_eiou_colname, r_eiou_colname, V_colname, Y_colname, S_units_colname}{Names of matrix columns.
See \code{IEATools::psut_cols}.}

\item{iea}{The string that identifies ECC data are from the IEA only.
Default is \code{PFUDatabase::ieamw_cols$iea}.}

\item{mw}{The string that identifies ECC data are for muscle work only.
Default is \code{PFUDatabase::ieamw_cols$mw}.}

\item{both}{The string that identifies ECC data are for both IEA and muscle work.
Default is \code{PFUDatabase::ieamw_cols}.
Default is \code{PFUDatabase::ieamw_cols$both}.}
}
\value{
A data frame with \code{PSUTIEA}, \code{PSUTMW}, and \code{PSUTIEAMW} \code{rbind()}ed together,
and a new column (\code{IEAMW_colname}) that distinguishes among them.
}
\description{
Combines PSUT descriptions based on IEA data exclusively, muscle work data exclusively,
and summed IEA and MW data.
}
