% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_iea_mw.R
\name{build_psut_dataframe}
\alias{build_psut_dataframe}
\title{Build the final PSUT data frame}
\usage{
build_psut_dataframe(
  PSUTIEA,
  PSUTMW,
  PSUTIEAMW,
  IEAMW_colname = PFUDatabase::ieamw_cols$ieamw,
  iea = PFUDatabase::ieamw_cols$iea,
  mw = PFUDatabase::ieamw_cols$mw,
  both = PFUDatabase::ieamw_cols$both,
  year = IEATools::iea_cols$year
)
}
\arguments{
\item{PSUTIEA}{A PSUT data frame of IEA data.}

\item{PSUTMW}{A PSUT data frame of muscle work data.}

\item{PSUTIEAMW}{A PSUT data frame of combined IEA and MW data.}

\item{IEAMW_colname}{The name of the column that identifies data source (IEA, MW, or both).
Default is \code{PFUDatabase::ieamw_cols$ieamw}.}

\item{iea}{The string that identifies ECC data are from the IEA only.
Default is \code{PFUDatabase::ieamw_cols$iea}.}

\item{mw}{The string that identifies ECC data are for muscle work only.
Default is \code{PFUDatabase::ieamw_cols$mw}.}

\item{both}{The string that identifies ECC data are for both IEA and muscle work.
Default is \code{PFUDatabase::ieamw_cols$both}.}
}
\value{
A data frame with \code{PSUTIEA}, \code{PSUTMW}, and \code{PSUTIEAMW} \code{rbind()}ed together,
and a new column (\code{IEAMW_colname}) that distinguishes among them.
}
\description{
Combines PSUT descriptions based on IEA data exclusively, muscle work data exclusively,
and summed IEA and MW data.
}
