% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exiobase.R
\name{calc_Ef_to_Xf_exiobase}
\alias{calc_Ef_to_Xf_exiobase}
\title{Calculates the final energy to final exergy multipliers}
\usage{
calc_Ef_to_Xf_exiobase(
  phi_vecs,
  years_exiobase,
  full_list_exiobase_flows,
  country_concordance_table_df
)
}
\arguments{
\item{phi_vecs}{A data frame of phi (exergy-to-energy ratio) coefficients.}

\item{years_exiobase}{The years for which the coefficients are provided to the Exiobase team}

\item{full_list_exiobase_flows}{The full list of energy flows used in the Exiobase pipeline led by KR}

\item{country_concordance_table_df}{A data frame containing the country concordance table}
}
\value{
A data frame of final energy to final exergy multipliers
}
\description{
Calculates the final energy to final exergy multipliers
}
