% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{is_balanced}
\alias{is_balanced}
\title{Tells whether IEA data are balanced}
\usage{
is_balanced(
  IEAData,
  countries,
  country = IEATools::iea_cols$country,
  grp_vars = c(country, IEATools::iea_cols$method, IEATools::iea_cols$energy_type,
    IEATools::iea_cols$last_stage, IEATools::iea_cols$year, IEATools::iea_cols$product)
)
}
\arguments{
\item{IEAData}{a tidy IEA data frame}

\item{countries}{the countries for which balancing should be checked as strings}

\item{country}{The name of the country column in \code{IEAData}. Default is Country.}

\item{grp_vars}{the groups that should be checked. Default is
\code{c(country, IEATools::iea_cols$method, IEATools::iea_cols$energy_type, IEATools::iea_cols$last_stage, IEATools::iea_cols$product)}.}
}
\value{
a logical stating whether all products are balanced for the country of interest
}
\description{
Performs the the energy balance check in a way that is amenable to drake subtargets.
Internally, this function uses \code{\link[IEATools:calc_tidy_iea_df_balances]{IEATools::calc_tidy_iea_df_balances()}}.
Grouping is doing internal to this function using the value of \code{grp_vars}.
}
\examples{
# These data are not balanced, because they are raw.
IEATools::sample_iea_data_path() \%>\%
  IEATools::load_tidy_iea_df() \%>\%
  is_balanced(countries = "ZAF")
}
