% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{machine_constants}
\alias{machine_constants}
\title{Information about the machine efficiency files}
\format{
A string list with 1 entries.
\describe{
\item{efficiency_tab_name}{The default name of the efficiency tabs in machine efficiency excel files.}
}
}
\usage{
machine_constants
}
\description{
A string list containing information about machine efficiency files.
Items in the list provide default values for machine efficiency files,
including Excel tab names, etc.
}
\examples{
machine_constants
}
\keyword{datasets}
