% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exiobase.R
\name{read_list_exiobase_energy_flows}
\alias{read_list_exiobase_energy_flows}
\title{Reads the list of Exiobase energy flows}
\usage{
read_list_exiobase_energy_flows(path_to_exiobase_data, list_energy_flows_file)
}
\arguments{
\item{path_to_exiobase_data}{Contains the path to the folder containing input data from Exiobase}

\item{list_energy_flows_file}{Contains the name of the file containing all the Exiobase energy flows}
}
\value{
A data frame containsing the list of Exiobase energy flows, and a boolean column stating whether the flow is a final energy flow or not
}
\description{
Reads the list of Exiobase energy flows
}
